+++ ruta-core/src/main/java/org/apache/uima/ruta/rule/RuleElement.java
@@ -61,7 +61,7 @@ public interface RuleElement {
 
   RuleElementQuantifier getQuantifier();
 
-  int estimateAnchors(RutaStream stream);
+  long estimateAnchors(RutaStream stream);
 
   List<Integer> getSelfIndexList();
